requirements.txt
fastapi==0.88.0
pydantic==1.10.2
python_dateutil==2.8.2
requests==2.28.1
 python-dotenv==0.21.0
